Move endpoint validation into the library
