+++ cmd/routedns/main.go
@@ -6,8 +6,6 @@ import (
 	"net"
 	"net/url"
 	"os"
-	"strconv"
-	"strings"
 	"time"
 
 	rdns "github.com/folbricht/routedns"
@@ -76,10 +74,6 @@ func start(opt options, args []string) error {
 			return fmt.Errorf("group resolver with duplicate id '%s", id)
 		}
 
-		if err := validEndpoint(r.Address); err != nil {
-			return err
-		}
-
 		switch r.Protocol {
 		case "doq":
 			tlsConfig, err := rdns.TLSClientConfig(r.CA, r.ClientCrt, r.ClientKey)
@@ -143,7 +137,10 @@ func start(opt options, args []string) error {
 			opt := rdns.DNSClientOptions{
 				LocalAddr: net.ParseIP(r.LocalAddr),
 			}
-			resolvers[id] = rdns.NewDNSClient(id, r.Address, r.Protocol, opt)
+			resolvers[id], err = rdns.NewDNSClient(id, r.Address, r.Protocol, opt)
+			if err != nil {
+				return fmt.Errorf("failed to parse resolver config for '%s' : %s", id, err)
+			}
 		default:
 			return fmt.Errorf("unsupported protocol '%s' for resolver '%s'", r.Protocol, id)
 		}
@@ -652,55 +649,4 @@ func newIPBlocklistDB(format, source, locationDB string, rules []string) (rdns.I
 	default:
 		return nil, fmt.Errorf("unsupported format '%s'", format)
 	}
-}
-
-// Returns nil if the endpoint address is a valid IP or name
-func validEndpoint(addr string) error {
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		return err
-	}
-	if _, err := strconv.ParseUint(port, 10, 16); err != nil {
-		return fmt.Errorf("invalid port: %w", err)
-	}
-	// See if we have a valid IP
-	if ip := net.ParseIP(host); ip != nil {
-		return nil
-	}
-	return validHostname(host)
-}
-
-// Returns nil if the given name is a valid hostnam as per https://tools.ietf.org/html/rfc3696#section-2
-// and https://tools.ietf.org/html/rfc1123#page-13
-func validHostname(name string) error {
-	if name == "" {
-		return errors.New("hostname empty")
-	}
-	if len(name) > 255 {
-		return fmt.Errorf("invalid hostname %q: too long", name)
-	}
-	name = strings.TrimSuffix(name, ".")
-	labels := strings.Split(name, ".")
-	for _, label := range labels {
-		for _, c := range label {
-			if label == "" {
-				return fmt.Errorf("invalid hostname %q: empty label", name)
-			}
-			if strings.HasPrefix(label, "-") || strings.HasSuffix(label, "-") {
-				return fmt.Errorf("invalid hostname %q: label can not start or end with -", name)
-			}
-			switch {
-			case c >= '0' && c <= '9', c >= 'a' && c <= 'z', c >= 'A' && c <= 'Z', c == '-':
-			default:
-				return fmt.Errorf("invalid hostname %q: invalid character %q", name, string(c))
-			}
-		}
-	}
-	// The last label can not be all-numeric
-	for _, c := range labels[len(labels)-1] {
-		if c < '0' || c > '9' {
-			return nil
-		}
-	}
-	return fmt.Errorf("invalid hostname %q: last label can not be all numeric", name)
 }
