+++ dnsclient.go
@@ -25,7 +25,10 @@ var _ Resolver = &DNSClient{}
 
 // NewDNSClient returns a new instance of DNSClient which is a plain DNS resolver
 // that supports pipelining over a single connection.
-func NewDNSClient(id, endpoint, network string, opt DNSClientOptions) *DNSClient {
+func NewDNSClient(id, endpoint, network string, opt DNSClientOptions) (*DNSClient, error) {
+	if err := validEndpoint(endpoint); err != nil {
+		return nil, err
+	}
 	// Use a custom dialer if a local address was provided
 	var dialer *net.Dialer
 	if opt.LocalAddr != nil {
@@ -47,7 +50,7 @@ func NewDNSClient(id, endpoint, network string, opt DNSClientOptions) *DNSClient
 		net:      network,
 		endpoint: endpoint,
 		pipeline: NewPipeline(endpoint, client),
-	}
+	}, nil
 }
 
 // Resolve a DNS query.
